+++ plugins/postcss-sign-functions/src/index.ts
@@ -12,6 +12,7 @@ export type pluginOptions = {
 
 const FUNCTION_CALL_REGEX = /(?<![-\w])(?:sign|abs)\(/i;
 const ABS_CALL_REGEX = /(?<![-\w])(?:sign|abs)\(/i;
+const IS_PROPERTY_REGEX = /^property$/i;
 
 const creator: PluginCreator<pluginOptions> = (opts?: pluginOptions) => {
 	const options: pluginOptions = Object.assign(
@@ -27,6 +28,10 @@ const creator: PluginCreator<pluginOptions> = (opts?: pluginOptions) => {
 		postcssPlugin: 'postcss-sign-functions',
 		Declaration(decl): void {
 			if (!FUNCTION_CALL_REGEX.test(decl.value)) {
+				return;
+			}
+
+			if (options.preserve && decl.parent?.type === 'atrule' && IS_PROPERTY_REGEX.test(decl.parent.name)) {
 				return;
 			}